Add frozen torch package
Use the same freezing functionality that we use for freezing the
standard library to freeze `torch`. Some hacks were necessary in places
where we use `__file__` in PyTorch. These need to be followed up on.

At this point, the unit tests pass even if we zero out `sys.path`
entirely, and strace confirms we are not accessing the file system for
anything (except to service linecache queries, which may also want to
follow up on).

Followups are commented as TODO(torchpy)