Add _get_type() API to RRef
The new API returns the type of the data object referenced by this
`RRef`. On the owner, this is same as `type(rref.local_value())`.
On a user, this will trigger an RPC to fetch the `type` object from
the owner. After this function is run once, the `type` object is
cached by the `RRef`, and subsequent invocations no longer trigger
RPC.

closes #33210

ghstack-source-id: 516beccbd379ba5766f9baa3aac424d1261d4c5e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44663